Adjust the size for the send buffer for frontend connections
The default is 4k, which is a bit excessive for how the buffer is used
(unless someone's using notify payloads, which they probably shouldn't
be).  128 bytes ought to be enough for a handful of normal notifications,
and it'll still bring the memory usage per connection down by almost 4
kilobytes.

It might make sense to make this configurable in the future, but
probably not.