Add a special hack for JDBC versions 9.1 to 9.3
These versions will try to set extra_float_digits after connecting if
they think the server version is 9.0 or higher, and get angry if they
can't.  Fix by accepting it as a "nop SET command".  This is safe to do
unconditionally, because we never actually handle floats (or real query
results, for that matter).