Implement a workaround for PostgreSQL bug #14830
The LISTEN / NOTIFY backend code didn't prepare for the possibility that
someone would have a long-lived listener backend which never executes a
query that requires a snapshot.  After some time (depending on how many
xids per second the cluster is doing) a significant portion of
notifications could easily be missed because of this.  This has been
fixed in postgres proper in commit
118e99c3d71efbea85341697a447d84bbfb54f18 which will be included in the
next back branch releases.  However, since a major user of allas is
still using PostgreSQL 9.1 and the workaround is rather simple and low
risk, I'm pushing this fix anyway.

The only really problematic part of this fix is that it requires hacking
pq.ListenerConn because the empty statement, when executed, does not
acquire a snapshot.  Unfortunate, but not really a big deal.